requirements.txt
torch==1.10.0
torchvision==0.11.1
onnx==1.10.0
onnxruntime==1.9.0
segmentation-models-pytorch==0.2.1
numpy==1.21.2
matplotlib==3.4.3
scikit-learn==0.24.2
Pillow==8.2.0
opencv-python==4.5.3.56
tqdm==4.62.3
pyyaml==5.4.1
scipy==1.7.1